pyproject.toml
[tool.poetry]
name = "notpixelbot"
version = "1.2.0"
description = "Bot for automation Not Pixel telegram mini app"
authors = ["Dellenoam", "YourLov3r"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
asyncio = "^3.4.3"
pydantic-settings = "^2.5.2"
argparse = "^1.4.0"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
better-proxy = "^1.2.0"
colorlog = "^6.8.2"
brotli = "^1.1.0"
pyjwt = "^2.9.0"
aiohttp-socks = "^0.9.0"
protobuf = "^5.28.3"
numpy = "^2.1.2"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"
typing-extensions = "^4.12.2"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
